Got existing tests to pass.
